+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Added
 
 - `limestone/Icon` supported icon list, adding new icon `f1`
+- `limestone/PageViews` `showFooterButtons` prop to show optional Close/Next buttons
+- `limestone/PageViews` `footerCloseLabel` prop to custom the Close button label
+- `limestone/PageViews` `onFooterCloseClick` prop to custom the Close button behavior when it is clicked
 - `limestone/Slider` `colorPicker` prop to change slider functionality
 - `limestone/ThemeDecorator` config `focusEffectClass` to support customization for spottable components